backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
sentence-transformers==3.0.1
chromadb==0.5.3
requests==2.32.3
python-dotenv==1.0.1
supabase==2.5.0
apify-client==1.7.1
beautifulsoup4==4.14.3
lxml==6.0.2
boto3==1.34.0
anthropic==0.18.0

# Testing dependencies
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.3
faker==22.6.0
locust==2.20.0
sentry-sdk==1.40.0
